requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
psycopg2-binary==2.9.9
bcrypt==4.1.2
PyJWT==2.8.0
pydantic==2.6.1
